Creating JavaFX project with Mavena and Github Actions
