+++ src/main/java/module-info.java
@@ -1,0 +1,6 @@
+module org.example {
+    requires javafx.controls;
+    requires javafx.fxml;
+    opens org.example to javafx.fxml;
+    exports org.example;
+}
