+++ src/main/java/org/example/FXMLController.java
@@ -1,0 +1,29 @@
+package org.example;
+/*
+Put header here
+
+
+ */
+
+import java.net.URL;
+import java.util.ResourceBundle;
+import javafx.event.ActionEvent;
+import javafx.fxml.FXML;
+import javafx.fxml.Initializable;
+import javafx.scene.control.Label;
+
+public class FXMLController implements Initializable {
+    
+    @FXML
+    private Label lblOut;
+    
+    @FXML
+    private void btnClickAction(ActionEvent event) {
+        lblOut.setText("Hello World!");
+    }
+    
+    @Override
+    public void initialize(URL url, ResourceBundle rb) {
+        // TODO
+    }    
+}
